Finished code review comments
