+++ doc/config/DoxygenLayout.xml
@@ -77,6 +77,7 @@
         <tab type="user" url="@ref reactive-infiltration" title="Reactive Infiltration"/>
       </tab>
       <tab type="user" url="@ref phase-field-fracture" title="Phase-Field Fracture"/>
+      <tab type="user" url="@ref finite-volume" title="Finite Volume Methods"/>
     </tab>
     <tab type="topics" title="Aditional Documentation"/>
     <tab type="usergroup" title="C++ API Reference">
