+++ doc/content/physics/finite_volume.md
@@ -2,9 +2,11 @@
 
 [TOC]
 
-The finite volume transport physics module provides composable building blocks for 1D finite-volume advection–diffusion–reaction systems. The models are designed for arbitrary batch dimensions and integrate with the standard NEML2 time integration and nonlinear solve infrastructure.
+The finite volume transport physics module provides composable building blocks for solve 1D PDEs with finite volume methods. The models are designed for arbitrary batch dimensions and 
+integrate with the standard NEML2 time integration and nonlinear solve infrastructure.  The module could be used for arbitrary PDEs, but the provided examples and tests focus
+on 1D transport (advection-diffusion-reaction) PDEs.
 
-## Governing equation
+## Finite volume discretization
 
 We consider a conserved quantity \f$u(x,t)\f$ with total flux \f$J\f$ and reaction term \f$R\f$:
 
@@ -12,23 +14,7 @@ We consider a conserved quantity \f$u(x,t)\f$ with total flux \f$J\f$ and reacti
   u_t + J_x = R,
 \f]
 
-with
-
-\f[
-  J = J_{diffusion} + J_{advection}, \quad
-  J_{diffusion} = -D u_x, \quad
-  J_{advection} = v u.
-\f]
-
-This yields the standard advection–diffusion–reaction equation
-
-\f[
-  u_t + (v u)_x - (D u_x)_x = R.
-\f]
-
-Boundary conditions can be Dirichlet or Neumann (flux) on the left/right ends of the domain.
-
-## Finite-volume discretization
+over a domain \f$\Omega \in [a,b]\f$.
 
 Partition \f$[a,b]\f$ into cells
 
@@ -49,6 +35,7 @@ Integrating the PDE over \f$I_i\f$ gives
   \frac{d \bar{u}_i}{dt} + \frac{1}{\Delta x_i}\left(J\big\rvert_{x_{i+\frac{1}{2}}} - J\big\rvert_{x_{i-\frac{1}{2}}}\right) = \bar{R}_i.
 \f]
 
+
 ### Cell-edge interpolation
 
 For nonuniform grids, cell-edge values are obtained by linear interpolation:
@@ -58,9 +45,9 @@ For nonuniform grids, cell-edge values are obtained by linear interpolation:
   + \frac{x_{i+\frac{1}{2}}-x_i}{x_{i+1}-x_i} q_{i+1}.
 \f]
 
-### Prefactor-weighted gradient
+### Gradients
 
-The prefactor-weighted gradient uses a first-order reconstructed gradient:
+Gradient (and in 1D divergence) terms are handled with the first order expression
 
 \f[
   \mathrm{grad}_{u, i+\frac{1}{2}} = -p_{i+\frac{1}{2}} \frac{\bar{u}_{i+1}-\bar{u}_i}{x_{i+1}-x_i}.
@@ -68,10 +55,10 @@ The prefactor-weighted gradient uses a first-order reconstructed gradient:
 
 ### Advective flux
 
-The advective flux uses first-order upwinding:
+The module provides an expression for stabilized advective fluxes via simple upwinding:
 
 \f[
-  \hat{J}_{advection, i+\frac{1}{2}} = v^+_{i+\frac{1}{2}} \bar{u}_i + v^-_{i+\frac{1}{2}} \bar{u}_{i+1},
+  \hat{J}_{\mathrm{advection}, i+\frac{1}{2}} = v^+_{i+\frac{1}{2}} \bar{u}_i + v^-_{i+\frac{1}{2}} \bar{u}_{i+1},
 \f]
 
 with
@@ -80,131 +67,28 @@ with
   v^\pm_{i+\frac{1}{2}} = \frac{1}{2}\left(v_{i+\frac{1}{2}} \pm |v_{i+\frac{1}{2}}|\right).
 \f]
 
+
 ## Finite volume transport model building blocks
 
 The finite volume transport module introduces several small models intended to be composed together:
 
 - **LinearlyInterpolateToCellEdges**: Interpolates cell-centered values onto cell edges on nonuniform grids.
 - **FiniteVolumeGradient**: Computes prefactor-weighted gradients at cell edges from \f$u\f$, edge prefactor values, and the spacing between cell centers.
-- **AdvectiveFlux**: Computes \f$J_{advection}\f$ at cell edges with first-order upwinding from \f$u\f$ and edge velocity \f$v_{edge}\f$.
-- **TransportBoundaryCondition**: Appends a boundary value to the left or right side of an intermediate dimension (useful for flux boundary conditions).
-- **ScalarLinearCombination**: Combines flux divergence and reaction to form \f$\dot{u}\f$.
+- **FiniteVolumeUpwindedAdvectiveFlux**: Computes \f$J_{\mathrm{advection}}\f$ at cell edges with first-order upwinding from \f$u\f$ and edge velocity \f$v_{edge}\f$.
+- **FiniteVolumeAppendBoundaryCondition**: Appends a boundary value to the left or right side of an intermediate dimension (useful for both Dirichlet and Neumann boundary conditions).
 
 In addition, helper tensors are provided for common 1D mesh setups:
 
 - **LinspaceScalar**: Uniform edge locations or time grids.
 - **CenterScalar**: Cell centers from edge locations.
 - **DifferenceScalar**: Cell sizes from edge locations.
-- **GaussianScalar**: Convenient Gaussian initial conditions.
+- **GaussianScalar**: Convenient for initializing Gaussian initial conditions.
 
 ## Example: combined advection–diffusion–reaction
 
 Below is a compact example that assembles a full transport system, applies boundary conditions, and advances in time using backward Euler. The full regression test can be found in tests/regression/finite_volume/combined/model.i.
 
-```
-[Tensors]
-  [edges]
-    type = LinspaceScalar
-    start = 0.0
-    end = 1.0
-    nstep = 201
-    dim = 0
-    group = 'intermediate'
-  []
-  [centers]
-    type = CenterScalar
-    points = 'edges'
-  []
-  [dx_centers]
-    type = DifferenceScalar
-    points = 'centers'
-  []
-  [dx]
-    type = DifferenceScalar
-    points = 'edges'
-  []
-  [ic]
-    type = GaussianScalar
-    points = 'centers'
-    width = 0.05
-    height = 1.0
-    center = 0.25
-  []
-[]
-
-[Models]
-  [diffusivity]
-    type = LinearlyInterpolateToCellEdges
-    cell_values = 0.5
-    cell_centers = 'centers'
-    cell_edges = 'edges'
-    edge_values = 'state/prefactor'
-  []
-  [advection_velocity]
-    type = LinearlyInterpolateToCellEdges
-    cell_values = 0.4
-    cell_centers = 'centers'
-    cell_edges = 'edges'
-    edge_values = 'state/v_edge'
-  []
-  [diffusive_flux]
-    type = FiniteVolumeGradient
-    u = 'state/concentration'
-    prefactor = 'state/prefactor'
-    dx = 'dx_centers'
-  []
-  [advective_flux]
-    type = AdvectiveFlux
-    u = 'state/concentration'
-    v_edge = 'state/v_edge'
-  []
-  [reaction]
-    type = ScalarLineCombination
-    from_var = 'state/concentration'
-    to_var = 'state/R'
-    coefficients = 0.05
-  []
-  [total_flux]
-    type = ScalarLinearCombination
-    from_var = 'state/grad_u state/J_advection'
-    to_var = 'state/J'
-    coefficients = '1 1'
-  []
-  [left_bc]
-    type = TransportBoundaryCondition
-    input = 'state/J'
-    bc_value = 0.0
-    side = 'left'
-  []
-  [right_bc]
-    type = TransportBoundaryCondition
-    input = 'state/J_with_bc_left'
-    bc_value = 0.0
-    side = 'right'
-  []
-  [flux_divergence]
-    type = FiniteVolumeGradient
-    u = 'state/J_with_bc_left_with_bc_right'
-    prefactor = 1
-    dx = 'dx'
-    grad_u = 'state/flux_div'
-  []
-  [rate_of_change]
-    type = ScalarLinearCombination
-    from_var = 'state/R state/flux_div'
-    to_var = 'state/concentration_rate'
-    coefficients = '1 1'
-  []
-  [integrate_u]
-    type = ScalarBackwardEulerTimeIntegration
-    variable = 'state/concentration'
-  []
-  [implicit_rate]
-    type = ComposedModel
-    models = 'diffusivity advection_velocity diffusive_flux advective_flux reaction total_flux left_bc right_bc flux_divergence rate_of_change integrate_u'
-  []
-[]
-```
+@list-input:../../../tests/regression/finite_volume/combined/model.i
 
 ## Verification and regression tests
