+++ include/neml2/models/finite_volume/FiniteVolumeAppendBoundaryCondition.h
@@ -24,30 +24,40 @@
 
 #pragma once
 
+#include "neml2/base/EnumSelection.h"
 #include "neml2/models/Model.h"
 
 namespace neml2
 {
 /**
- * @brief First-order upwind advective flux at cell edges.
+ * @brief Append a boundary condition value to the intermediate dimension.
  */
-class AdvectiveFlux : public Model
+class FiniteVolumeAppendBoundaryCondition : public Model
 {
 public:
+  enum class Side : std::uint8_t
+  {
+    LEFT = 0,
+    RIGHT = 1
+  };
+
   static OptionSet expected_options();
 
-  AdvectiveFlux(const OptionSet & options);
+  FiniteVolumeAppendBoundaryCondition(const OptionSet & options);
 
 protected:
   void set_value(bool out, bool dout_din, bool d2out_din2) override;
 
-  /// Cell-averaged field values
-  const Variable<Scalar> & _u;
+  /// Input tensor (e.g., cell values or cell-edge fluxes)
+  const Variable<Scalar> & _input;
 
-  /// Cell-edge advection velocity values
-  const Variable<Scalar> & _v_edge;
+  /// Boundary value to append
+  const Scalar & _bc_value;
 
-  /// Cell-edge advective fluxes
-  Variable<Scalar> & _J;
+  /// Which side to append to
+  Side _side;
+
+  /// Output tensor with boundary condition appended
+  Variable<Scalar> & _output;
 };
 } // namespace neml2
