+++ tests/unit/models/finite_volume/FiniteVolumeAppendBoundaryCondition.i
@@ -22,8 +22,10 @@
     intermediate_dimension = 1
   []
   [u_bc]
-    type = FullScalar
-    value = 10
+    type = Scalar
+    values = '10'
+    batch_shape = '(1)'
+    intermediate_dimension = 1
   []
   [u_bar_with_bc_left]
     type = Scalar
@@ -35,7 +37,7 @@
 
 [Models]
   [model]
-    type = TransportBoundaryCondition
+    type = FiniteVolumeAppendBoundaryCondition
     input = 'state/u_bar'
     bc_value = 'u_bc'
     side = 'left'
