+++ tests/unit/models/finite_volume/FiniteVolumeAppendBoundaryCondition_batched_right.i
@@ -17,25 +17,27 @@
 [Tensors]
   [u_bar]
     type = Scalar
-    values = '1 2 3'
-    batch_shape = '(3)'
+    values = '1 2 3 4 5 6'
+    batch_shape = '(2,3)'
     intermediate_dimension = 1
   []
   [u_bc]
-    type = FullScalar
-    value = 10
+    type = Scalar
+    values = '10 20'
+    batch_shape = '(2,1)'
+    intermediate_dimension = 1
   []
   [u_bar_with_bc_right]
     type = Scalar
-    values = '1 2 3 10'
-    batch_shape = '(4)'
+    values = '1 2 3 10 4 5 6 20'
+    batch_shape = '(2,4)'
     intermediate_dimension = 1
   []
 []
 
 [Models]
   [model]
-    type = TransportBoundaryCondition
+    type = FiniteVolumeAppendBoundaryCondition
     input = 'state/u_bar'
     bc_value = 'u_bc'
     side = 'right'
