+++ tests/unit/models/finite_volume/FiniteVolumeAppendBoundaryCondition_right.i
@@ -5,11 +5,11 @@
     check_AD_parameter_derivatives = false
     input_Scalar_names = 'state/u_bar'
     input_Scalar_values = 'u_bar'
-    output_Scalar_names = 'state/u_bar_with_bc_right'
-    output_Scalar_values = 'u_bar_with_bc_right'
+    output_Scalar_names = 'state/u_bar_bc'
+    output_Scalar_values = 'u_bar_bc'
     input_with_intrsc_intmd_dims = 'state/u_bar'
     input_intrsc_intmd_dims = '1'
-    output_with_intrsc_intmd_dims = 'state/u_bar_with_bc_right'
+    output_with_intrsc_intmd_dims = 'state/u_bar_bc'
     output_intrsc_intmd_dims = '1'
   []
 []
@@ -17,29 +17,36 @@
 [Tensors]
   [u_bar]
     type = Scalar
-    values = '1 2 3 4 5 6'
-    batch_shape = '(2,3)'
+    values = '1 2 3'
+    batch_shape = '(3)'
     intermediate_dimension = 1
   []
   [u_bc]
     type = Scalar
-    values = '10 20'
-    batch_shape = '(2,1)'
+    values = '10'
+    batch_shape = '(1)'
     intermediate_dimension = 1
   []
-  [u_bar_with_bc_right]
+  [u_bar_with_bc]
     type = Scalar
-    values = '1 2 3 10 4 5 6 20'
-    batch_shape = '(2,4)'
+    values = '1 2 3 10'
+    batch_shape = '(4)'
+    intermediate_dimension = 1
+  []
+  [u_bar_bc]
+    type = Scalar
+    values = '1 2 3 10'
+    batch_shape = '(4)'
     intermediate_dimension = 1
   []
 []
 
 [Models]
   [model]
-    type = TransportBoundaryCondition
+    type = FiniteVolumeAppendBoundaryCondition
     input = 'state/u_bar'
     bc_value = 'u_bc'
     side = 'right'
+    output = 'state/u_bar_bc'
   []
 []
