+++ tests/unit/models/finite_volume/FiniteVolumeAppendBoundaryCondition_scalar_bc_left.i
@@ -5,11 +5,11 @@
     check_AD_parameter_derivatives = false
     input_Scalar_names = 'state/u_bar'
     input_Scalar_values = 'u_bar'
-    output_Scalar_names = 'state/u_bar_bc'
-    output_Scalar_values = 'u_bar_bc'
+    output_Scalar_names = 'state/u_bar_with_bc_left'
+    output_Scalar_values = 'u_bar_with_bc_left'
     input_with_intrsc_intmd_dims = 'state/u_bar'
     input_intrsc_intmd_dims = '1'
-    output_with_intrsc_intmd_dims = 'state/u_bar_bc'
+    output_with_intrsc_intmd_dims = 'state/u_bar_with_bc_left'
     output_intrsc_intmd_dims = '1'
   []
 []
@@ -22,20 +22,12 @@
     intermediate_dimension = 1
   []
   [u_bc]
-    type = Scalar
-    values = '10'
-    batch_shape = '(1)'
-    intermediate_dimension = 1
+    type = FullScalar
+    value = 10
   []
-  [u_bar_with_bc]
+  [u_bar_with_bc_left]
     type = Scalar
-    values = '1 2 3 10'
-    batch_shape = '(4)'
-    intermediate_dimension = 1
-  []
-  [u_bar_bc]
-    type = Scalar
-    values = '1 2 3 10'
+    values = '10 1 2 3'
     batch_shape = '(4)'
     intermediate_dimension = 1
   []
@@ -43,10 +35,9 @@
 
 [Models]
   [model]
-    type = TransportBoundaryCondition
+    type = FiniteVolumeAppendBoundaryCondition
     input = 'state/u_bar'
     bc_value = 'u_bc'
-    side = 'right'
-    output = 'state/u_bar_bc'
+    side = 'left'
   []
 []
