+++ tests/unit/models/finite_volume/FiniteVolumeUpwindedAdvectiveFlux_scalar_v.i
@@ -35,7 +35,7 @@
 
 [Models]
   [model]
-    type = AdvectiveFlux
+    type = FiniteVolumeUpwindedAdvectiveFlux
     u = 'state/u'
     v_edge = 'state/v_edge'
     flux = 'state/J'
