+++ tests/verification/finite_volume/advection/advection.i
@@ -122,7 +122,7 @@ final_center = 0.75
       dx = 'dx_centers'
   []
   [advective_flux]
-      type = AdvectiveFlux
+      type = FiniteVolumeUpwindedAdvectiveFlux
       u = 'state/concentration'
       v_edge = 'state/v_edge'
   []
@@ -139,13 +139,13 @@ final_center = 0.75
     coefficients = '1 1'
   []
   [left_bc]
-    type = TransportBoundaryCondition
+    type = FiniteVolumeAppendBoundaryCondition
     input = 'state/J'
     bc_value = 0.0
     side = 'left'
   []
   [right_bc]
-    type = TransportBoundaryCondition
+    type = FiniteVolumeAppendBoundaryCondition
     input = 'state/J_with_bc_left'
     bc_value = 0.0
     side = 'right'
